Remove logic which traps libpostal parsing errors
The new libpostal v.1 parses well and does not push messed up items
to wrong query components. We can use libostal parsing as it is,
except that we fallback to addressit if libpostal produces a single
'query' string. In such cases, addressit puts the items after comma
separator to admin regions which is the right way.